+++ lib/structure/navigation.php
@@ -53,7 +53,47 @@ remove_action( 'genesis_after_endwhile', 'genesis_posts_nav' );
 add_action( 'genesis_after_loop', 'genesis_posts_nav' );
 
 // Add previous/next post links to single posts
-add_action( 'genesis_entry_footer', 'genesis_prev_next_post_nav' );
+// add_action( 'genesis_entry_footer', 'genesis_prev_next_post_nav' );
+
+add_post_type_support( 'post', 'genesis-adjacent-entry-nav' );
+
+/**
+ * Filters the adjacent post link.
+ *
+ * The dynamic portion of the hook name, `$adjacent`, refers to the type
+ * of adjacency, 'next' or 'previous'.
+ *
+ * @since 2.6.0
+ * @since 4.2.0 Added the `$adjacent` parameter.
+ *
+ * @param string  $output   The adjacent post link.
+ * @param string  $format   Link anchor format.
+ * @param string  $link     Link permalink format.
+ * @param WP_Post $post     The adjacent post.
+ * @param string  $adjacent Whether the post is previous or next.
+ */
+add_filter( 'previous_post_link', 'mai_adjacent_post_link_thumbnail', 10, 5 );
+add_filter( 'next_post_link', 'mai_adjacent_post_link_thumbnail', 10, 5 );
+function mai_adjacent_post_link_thumbnail( $output, $format, $link, $post, $adjacent ) {
+	$output = _mai_get_adjacent_post_output( $output, $post );
+	$output = str_replace( '&#x000AB;', '<span class="pagination-icon">&#x000AB;</span>', $output );
+	$output = str_replace( '&#x000BB;', '<span class="pagination-icon">&#x000BB;</span>', $output );
+	return $output;
+}
+
+function _mai_get_adjacent_post_output( $output, $post ) {
+	$image_id = get_post_thumbnail_id( $post );
+	if ( $image_id ) {
+		$image = wp_get_attachment_image( $image_id, 'tiny' );
+		if ( $image ) {
+			// Previous
+			$output = str_replace( '&#x000AB;', '&#x000AB;' . $image, $output );
+			// Next
+			$output = str_replace( '&#x000BB;', $image . '&#x000BB;', $output );
+		}
+	}
+	return $output;
+}
 
 /**
  * Add mobile menu on custom hook, after the site header row